requirements.txt
Django==5.2.6
pillow==11.3.0
django-summernote==0.8.20.0
django-unfold==0.65.0